refactor: seperate calendar module
