+++ escahp/src/main/java/com/esca/escahp/calendar/CalendarController.java
@@ -1,0 +1,8 @@
+package com.esca.escahp.calendar;
+
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class CalendarController {
+}
