+++ escahp/src/main/java/com/esca/escahp/calendar/CalendarService.java
@@ -1,0 +1,7 @@
+package com.esca.escahp.calendar;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class CalendarService {
+}
