+++ escahp/src/main/java/com/esca/escahp/calendar/I_CalendarService.java
@@ -1,0 +1,4 @@
+package com.esca.escahp.calendar;
+
+public interface I_CalendarService {
+}
